@@ -15,32 +15,33 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.1.0
+version: 0.2.0
 
 environment:
-  sdk: ">=2.10.0-0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
   path: ^1.8.0
-  sembast: ^3.0.0+4
-  font_awesome_flutter: ^8.11.0
+  sembast: ^3.0.2
+  font_awesome_flutter: ^9.0.0
   built_collection: ^5.0.0
-  built_value: ^8.0.0
-  provider: ^4.3.3
+  built_value: ^8.0.6
+  provider: ^5.0.0
   flutter:
     sdk: flutter
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  google_fonts: ^1.1.2
+  cupertino_icons: ^1.0.3
+  google_fonts: ^2.0.0
   path_provider: ^2.0.1
   flutter_slidable: ^0.6.0
+  animations: ^2.0.0
 
 dev_dependencies:
-  build_runner: ^1.0.0
-  built_value_generator: ^8.0.0
-  pedantic: ^1.4.0
-  quiver: '>=0.21.0 <3.0.0'
-  test: ^1.0.0
+  build_runner: ^2.0.2
+  built_value_generator: ^8.0.6
+  pedantic: ^1.11.0
+  quiver: ^3.0.1
+  test: ^1.16.5
   flutter_test:
     sdk: flutter
 
