@@ -11,7 +11,7 @@ class AppDatabase {
   static final AppDatabase _singleton = AppDatabase._();
 
   static AppDatabase get instance => _singleton;
-  Completer<Database> _dbOpenCompleter;
+  Completer<Database>? _dbOpenCompleter;
 
   AppDatabase._();
 
@@ -20,15 +20,15 @@ class AppDatabase {
       _dbOpenCompleter = Completer();
       _openDatabase();
     }
-    return _dbOpenCompleter.future;
+    return _dbOpenCompleter!.future;
   }
 
   Future _openDatabase() async {
     final appDocumentDir = await getApplicationDocumentsDirectory();
-    final dbPath = appDocumentDir.path + DB_NAME;
+    final dbPath = appDocumentDir.path + '/' + DB_NAME;
     print(dbPath);
 
     final database = await databaseFactoryIo.openDatabase(dbPath);
-    _dbOpenCompleter.complete(database);
+    _dbOpenCompleter!.complete(database);
   }
 }
